define find_files() to return a simple list
Previously, `find_files()` would return a weird 2-tuple of the directory and a
list of any files it contained, resembling the output of `os.walk()`. To improve
readability and remove an excessive second `for` loop in main, `find_files()`
was re-written to only return a list of the absolute paths of each file.

In addition, `os.mkdir()` will no longer throw exceptions, as `normalizer()` now
checks to see if the output directory exists before making it. This was part of
the re-write of `normalizer()` to support the changes in `find_files()`.

Finally, trailing whitespace was also removed.