remove verbose option, update README.md to reflect
Upon further reflection, there's no need to hide output data from the user. The
user should be able to attempt debugging my software in the situation something
goes wrong, regardless of whether he is using a terminal or an IDE unable of
providing command-line flags. I took a little inspiration from the `mpv` project
for this, as they only have a quiet option instead of a verbose option.